moved tickets.py to ticket.py + some typing fixes
